requirements.txt
django~=4.1
djangorestframework~=3.14.0
Pillow~=9.3.0
psycopg2-binary~=2.9.5
drf-yasg~=1.21.4
django-admin-interface~=0.24.2
django-sortedm2m~=3.1.1
django-admin-sortable2~=2.1.4
django-object-actions~=4.1.0
django-nested-admin~=4.0.2
